+++ backend/src/main/java/com/voiceprint/backend/api/thema/dto/DiaryThemaCreateResponse.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class DiaryThemaCreateResponse {
-    private Long themaId;
+    private Integer themaId;
     private String example;
 }
